Fix migration for CI and initial migration

+++ db/migrate/20210601204622_add_index_for_better_show_model_year_budget_widget_search.rb
@@ -2,10 +2,10 @@
 class AddIndexForBetterShowModelYearBudgetWidgetSearch < ActiveRecord::Migration[6.1]
   def up
     ids = execute("SELECT id FROM ad_option_types WHERE name IN ('maker', 'model', 'year')").to_a.map { |x| x['id'] }
-    add_index(:ad_options, :ad_option_value_id, where: "ad_option_type_id IN (#{ids.join(',')})")
+    add_index(:ad_options, :ad_option_value_id, where: "ad_option_type_id IN (#{ids.join(',')})") if ids.present?
   end
 
   def down
-    remove_index(:ad_options, :ad_option_value_id)
+    remove_index(:ad_options, :ad_option_value_id) if index_exists?(:ad_options, :ad_option_value_id)
   end
 end
